chore(ptokens.js): updates test for enclave and peos
